Title Scroll updates
Added more to the story itself, created a new source file for the scroll font to change the size, and changed the project settings so that when the window is resized everything resizes. Still need to figure out how to center the text and make it shrink like the star wars text scroll.